+++ src/emulator/registry.ts
@@ -2,9 +2,9 @@ import * as clc from "cli-color";
 
 import { ALL_EMULATORS, EmulatorInstance, Emulators, EmulatorInfo } from "./types";
 import { FirebaseError } from "../error";
-import * as utils from "../utils";
 import * as controller from "./controller";
 import { Constants } from "./constants";
+import { EmulatorLogger } from "./emulatorLogger";
 
 /**
  * Static registry for running emulators to discover each other.
@@ -25,11 +25,6 @@ export class EmulatorRegistry {
     const info = instance.getInfo();
     await controller.waitForPortClosed(info.port, info.host);
     this.set(instance.getName(), instance);
-
-    utils.logLabeledSuccess(
-      instance.getName(),
-      `${description} started at ${clc.bold.underline(`http://${info.host}:${info.port}`)}`
-    );
   }
 
   static async stop(name: Emulators): Promise<void> {