+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 - Adds new API commands that handle list/create/delete operations on the Android SHA certificate hashes `apps:android:sha:list`, `apps:android:sha:create`, and `apps:android:sha:delete`.
 - Fixes an issue where the CLI did not assume admin privileges when performing Firestore / RTDB emulator operations.
 - Fixes an issue where the functions and hosting emulators would crash when not properly initialized (#2112).
+- Improves logs when running `emulators:start`.
 - Fixes an issue where `use` would allow an invalid project identifier.
 - Fixes an issue where custom options passed to `admin.initializeApp()` in the functions emulator were improperly augmented.
 - Changes `firebasemods.*.*` IAM permission checks to `firebaseextensions.*.*`
 - Fixes an issue where the suggested redeploy command for Firebase Functions was incorrect for names with dashes.
 - Fixes an issue where unicode characters in auth tokens crashed callable auth (#2185).
+- Fixes an issue where emulators could not be cleanly shut down if they had not started properly (#2228).